% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop}
\alias{chop}
\alias{chop.default}
\alias{chop.numeric}
\alias{chop.data.frame}
\title{Convert a continuous variable to a discrete variable.}
\usage{
chop(x, ...)

\method{chop}{default}(x, ...)

\method{chop}{numeric}(x, ...)

\method{chop}{data.frame}(x, .factorVars, .skip, .m, .g, ...)
}
\arguments{
\item{x}{A numeric or integer vector.}

\item{...}{Additional arguments passed to \code{hist()} or \code{Hmisc::cut2()}}

\item{.m}{(optional) Default value for the 'm' parameter to \code{Hmisc::cut2()}}

\item{.g}{(optional) Default value for the 'g' parameter to \code{Hmisc::cut2()}}

\item{factorVars}{(optional) A character vector of variable names in \code{x} to be directly converted to factors.}

\item{skip}{(optional) A character vector of variable names in \code{x} to skip processing.}
}
\value{
A factor the same length as \code{x}.
}
\description{
Convert a continuous variable to a discrete variable.
}
\details{
When \code{x} is numeric, \code{chop()} assumes that any specified values for \code{cuts},
\code{g}, or \code{m} are exactly as they are required to be in \code{cut2()}. When \code{x} is a
data frame, on the other hand, these parameters (and any other parameters passed to \code{cut2()}
or \code{hist()})  are specified through parameters that have names which are common to the names of
numeric varibles in \code{x}. (Any such specifications for non-numeric variables in \code{x} are
ignored.) See examples for a demonstration of this usage. Note that \code{.m} and \code{.g} are
overridden by any explicit values given for specific variables.  See behavior for variable 'drat'
in the final example.
}
\examples{
table(chop(0:100))
table(chop(0:100000))
table(chop(0:100, breaks = c(-Inf,23,68,Inf))) #breaks passed to hist()
table(chop(0:100, cuts   = c(-Inf,23,68,Inf))) #same result, but cuts is passed to cut2()
table(chop(0:100, g=4))
table(chop(0:100, m=17))
summary(chop(mtcars, .m=10, .factorVars = c("cyl","vs","am","gear")))
z <-  chop(mtcars,
           .factorVars = c("cyl","vs","am","gear"),
           .skip = c("mpg"),
           .g = 4,
           drat = list(g=8),
           qsec = list(cuts=c(0,16,18,20,100)))
summary(z)
table(z$drat)

g_chop <- chop(glucose, glucose = list(cuts = c(0,109,125,300)))
table(g_chop$glucose, g_chop$cohort)
}
\seealso{
\code{\link{hist}} and \code{\link{Hmisc::cut2}}
for more details.
}
