% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/divergence.R
\name{JSD}
\alias{JSD}
\title{Computes the Jensen-Shannon divergence between two probabiliy distributions.}
\usage{
JSD(P, Q)
}
\arguments{
\item{P}{A probility distribution (vector summing to one).}

\item{Q}{A probility distribution (vector summing to one).}
}
\value{
The JSD of \code{P} and \code{Q}.
}
\description{
Computes the Jensen-Shannon divergence between two probabiliy distributions.
}
\examples{
P = prop.table(sample(1:10, 20, replace = TRUE))
Q = prop.table(sample(5:15, 20, replace = TRUE))

JSD(P,Q)
JSD(Q,P)
}

