% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/divergence.R
\name{KLD}
\alias{KLD}
\title{Computes the Kullback-Leibler divergence between two probabiliy distributions.}
\usage{
KLD(A, B)
}
\arguments{
\item{A}{A probility distribution (vector summing to one).}

\item{B}{A probility distribution (vector summing to one).}
}
\value{
The KLD of \code{A} and \code{B}.
}
\description{

}
\details{
KLD is defined as: \deqn{KLD(A,B) = \sum{A \times \log\left(\frac{A}{B}\right)}}
}
\examples{
A = prop.table(sample(1:10, 20, replace = TRUE))
B = prop.table(sample(5:15, 20, replace = TRUE))

KLD(A,B)
KLD(B,A)
}

